+++ packages/graphiql/resources/webpack.config.js
@@ -19,9 +19,9 @@ const resultConfig = {
         'react-hot-loader/patch', // activate HMR for React
         'webpack-dev-server/client?http://localhost:8080', // bundle the client for webpack-dev-server and connect to the provided endpoint
         'webpack/hot/only-dev-server', // bundle the client for hot reloading, only- means to only hot reload for successful updates
-        './cdn.js', // the entry point of our app
+        './cdn.ts', // the entry point of our app
       ]
-    : './cdn.js',
+    : './cdn.ts',
   context: rootPath('src'),
   output: {
     path: rootPath(),
@@ -60,7 +60,7 @@ const resultConfig = {
       {
         test: /\.(js|jsx|ts|tsx)$/,
         use: [{ loader: 'babel-loader' }],
-        exclude: /\.(d\.ts|d\.ts\.map)$/,
+        exclude: /\.(d\.ts|d\.ts\.map|spec\.tsx)$/,
       },
       {
         test: /\.css$/,