+++ packages/graphql-language-service-server/src/startServer.ts
@@ -7,8 +7,9 @@
  *
  */
 
-import net from 'net';
+import * as net from 'net';
 import { MessageProcessor } from './MessageProcessor';
+import { GraphQLConfig } from 'graphql-config';
 
 import {
   createMessageConnection,
@@ -35,9 +36,13 @@ import {
   PublishDiagnosticsNotification,
   DidChangeWatchedFilesNotification,
   ShutdownRequest,
+  DocumentSymbolRequest,
+  // WorkspaceSymbolRequest,
+  // ReferencesRequest,
 } from 'vscode-languageserver';
 
 import { Logger } from './Logger';
+import { parseDocument } from './parseDocument';
 
 type Options = {
   // port for the LSP server to run on
@@ -46,7 +51,14 @@ type Options = {
   method?: 'socket' | 'stream' | 'node';
   // the directory where graphql-config is found
   configDir?: string;
+  // array of functions to transform the graphql-config and add extensions dynamically
+  extensions?: Array<(config: GraphQLConfig) => GraphQLConfig>;
+  fileExtensions?: string[];
+  // pre-existing GraphQLConfig
+  config?: GraphQLConfig;
+  parser?: typeof parseDocument;
 };
+('graphql-language-service-types');
 
 /**
  * startServer - initialize LSP server with options
@@ -54,7 +66,7 @@ type Options = {
  * @param options {Options} server initialization methods
  * @returns {Promise<void>}
  */
-export default (async function startServer(options: Options): Promise<void> {
+export default async function startServer(options: Options): Promise<void> {
   const logger = new Logger();
 
   if (options && options.method) {
@@ -83,7 +95,13 @@ export default (async function startServer(options: Options): Promise<void> {
               process.exit(0);
             });
             const connection = createMessageConnection(reader, writer, logger);
-            addHandlers(connection, logger, options.configDir);
+            addHandlers(
+              connection,
+              logger,
+              options.configDir,
+              options?.extensions ?? [],
+              options.config,
+            );
             connection.listen();
           })
           .listen(port);
@@ -99,17 +117,35 @@ export default (async function startServer(options: Options): Promise<void> {
         break;
     }
     const connection = createMessageConnection(reader, writer, logger);
-    addHandlers(connection, logger, options.configDir);
+    addHandlers(
+      connection,
+      logger,
+      options.configDir,
+      options?.extensions ?? [],
+      options.config,
+      options.parser,
+      options.fileExtensions,
+    );
     connection.listen();
   }
-});
+}
 
 function addHandlers(
   connection: MessageConnection,
   logger: Logger,
   configDir?: string,
+  extensions?: Array<(config: GraphQLConfig) => GraphQLConfig>,
+  config?: GraphQLConfig,
+  parser?: typeof parseDocument,
+  fileExtensions?: string[],
 ): void {
-  const messageProcessor = new MessageProcessor(logger);
+  const messageProcessor = new MessageProcessor(
+    logger,
+    extensions,
+    config,
+    parser,
+    fileExtensions,
+  );
   connection.onNotification(
     DidOpenTextDocumentNotification.type,
     async params => {
@@ -182,4 +218,13 @@ function addHandlers(
   connection.onNotification(DidChangeWatchedFilesNotification.type, params =>
     messageProcessor.handleWatchedFilesChangedNotification(params),
   );
+  connection.onRequest(DocumentSymbolRequest.type, params =>
+    messageProcessor.handleDocumentSymbolRequest(params),
+  );
+  // connection.onRequest(WorkspaceSymbolRequest.type, params =>
+  //   messageProcessor.handleWorkspaceSymbolRequest(params),
+  // );
+  // connection.onRequest(ReferencesRequest.type, params =>
+  //   messageProcessor.handleReferencesRequest(params),
+  // );
 }